+++ src/main.py
@@ -1,260 +1,9 @@
-from dataclasses import dataclass
+import processor as p
+import entity as e
+
 import esper
 import tcod
-import sys
 
-
-# -----------------------------------------------------------------------------
-# Global Constants
-# -----------------------------------------------------------------------------
-
-SCREEN_WIDTH = 80
-SCREEN_HEIGHT = 50
-TITLE = 'libtcod tutorial revised'
-
-MAP_WIDTH = 80
-MAP_HEIGHT = 43
-
-FONT_PATH = 'data/consolas10x10.png'
-FONT_FLAG = tcod.FONT_TYPE_GREYSCALE | tcod.FONT_LAYOUT_TCOD
-
-
-# -----------------------------------------------------------------------------
-# Components
-# -----------------------------------------------------------------------------
-
-@dataclass
-class Event:
-    action: dict = None
-
-
-@dataclass
-class Position:
-    x: int
-    y: int
-
-
-@dataclass
-class Renderable:
-    char: str = '@'
-    fg: tuple = (255, 255, 255)
-    bg: tuple = (0, 0, 0)
-    bg_blend: int = tcod.BKGND_NONE
-
-
-# -----------------------------------------------------------------------------
-# Event Processor
-# -----------------------------------------------------------------------------
-
-@dataclass
-class Key:
-    vk: int = 65
-    ch: str = chr(0)
-    pressed: bool = True
-    alt: bool = False
-    ctrl: bool = False
-    meta: bool = False
-    shift: bool = False
-
-    def __key(self):
-        return self.vk, self.ch, self.pressed, self.alt, self.ctrl, self.shift
-
-    def __eq__(self, other):
-        return self.__key() == other.__key()
-
-    def __hash__(self):
-        return hash(self.__key())
-
-
-class EventProcessor(esper.Processor):
-    def __init__(self):
-        super().__init__()
-        self.key = tcod.Key()
-        self.mouse = tcod.Mouse()
-        self.key_code = {
-            Key(vk=tcod.KEY_ENTER, ch='\r', alt=True): {'fullscreen': True},
-            Key(vk=tcod.KEY_ESCAPE, ch='\x1b'): {'exit': True},
-            Key(vk=tcod.KEY_LEFT, shift=True): {'move': (-1, -1)},
-            Key(vk=tcod.KEY_RIGHT, shift=True): {'move': (1, -1)},
-            Key(vk=tcod.KEY_LEFT, ctrl=True): {'move': (-1, 1)},
-            Key(vk=tcod.KEY_RIGHT, ctrl=True): {'move': (1, 1)},
-            Key(vk=tcod.KEY_UP): {'move': (0, -1)},
-            Key(vk=tcod.KEY_DOWN): {'move': (0, 1)},
-            Key(vk=tcod.KEY_LEFT): {'move': (-1, 0)},
-            Key(vk=tcod.KEY_RIGHT): {'move': (1, 0)},
-            Key(vk=tcod.KEY_KP0): {'move': (0, 0)},
-            Key(vk=tcod.KEY_KP1): {'move': (-1, 1)},
-            Key(vk=tcod.KEY_KP2): {'move': (0, 1)},
-            Key(vk=tcod.KEY_KP3): {'move': (1, 1)},
-            Key(vk=tcod.KEY_KP4): {'move': (-1, 0)},
-            Key(vk=tcod.KEY_KP5): {'move': (0, 0)},
-            Key(vk=tcod.KEY_KP6): {'move': (1, 0)},
-            Key(vk=tcod.KEY_KP7): {'move': (-1, -1)},
-            Key(vk=tcod.KEY_KP8): {'move': (0, -1)},
-            Key(vk=tcod.KEY_KP9): {'move': (1, -1)},
-            Key(ch='a'): {},
-            Key(ch='b'): {'move': (-1, 1)},
-            Key(ch='c'): {},
-            Key(ch='d'): {},
-            Key(ch='e'): {},
-            Key(ch='f'): {},
-            Key(ch='g'): {},
-            Key(ch='h'): {'move': (-1, 0)},
-            Key(ch='i'): {},
-            Key(ch='j'): {'move': (0, 1)},
-            Key(ch='k'): {'move': (0, -1)},
-            Key(ch='l'): {'move': (1, 0)},
-            Key(ch='m'): {},
-            Key(ch='n'): {'move': (1, 1)},
-            Key(ch='o'): {},
-            Key(ch='p'): {},
-            Key(ch='q'): {},
-            Key(ch='r'): {},
-            Key(ch='s'): {},
-            Key(ch='t'): {},
-            Key(ch='u'): {'move': (1, -1)},
-            Key(ch='v'): {},
-            Key(ch='w'): {},
-            Key(ch='x'): {},
-            Key(ch='y'): {'move': (-1, -1)},
-            Key(ch='z'): {},
-            Key(ch='.'): {'move': (0, 0)}
-        }
-
-    def process(self):
-        for ent, event in self.world.get_component(Event):
-            tcod.sys_wait_for_event(
-                mask=tcod.EVENT_ANY,
-                k=self.key,
-                m=self.mouse,
-                flush=False
-            )
-
-            user_input = Key(
-                vk=self.key.vk, ch=chr(self.key.c),
-                alt=(self.key.lalt or self.key.ralt),
-                ctrl=(self.key.lctrl or self.key.lctrl),
-                meta=(self.key.lmeta or self.key.rmeta),
-                shift=self.key.shift, pressed=self.key.pressed,
-            )
-
-            if tcod.EVENT_KEY and user_input in self.key_code:
-                event.action = self.key_code[user_input]
-            else:
-                event.action = {}
-
-
-# -----------------------------------------------------------------------------
-# Movement Processor
-# -----------------------------------------------------------------------------
-
-class MovementProcessor(esper.Processor):
-    def __init__(self):
-        super().__init__()
-
-    def process(self):
-        for ent, (pos, event) in self.world.get_components(Position, Event):
-            move = event.action.get('move')
-            if move:
-                dx, dy = move
-                pos.x += dx
-                pos.y += dy
-
-
-# -----------------------------------------------------------------------------
-# Render Processor
-# -----------------------------------------------------------------------------
-
-class RenderProcessor(esper.Processor):
-    def __init__(self):
-        super().__init__()
-
-        self.screen_width = SCREEN_WIDTH
-        self.screen_height = SCREEN_HEIGHT
-        self.map_width = MAP_WIDTH
-        self.map_height = MAP_HEIGHT
-
-        tcod.console_set_custom_font(
-            fontFile=FONT_PATH,
-            flags=FONT_FLAG
-        )
-
-        self.root_console = tcod.console_init_root(
-            w=self.screen_width,
-            h=self.screen_height,
-            title=TITLE
-        )
-
-        self.con = tcod.console.Console(
-            width=self.map_width,
-            height=self.map_height
-        )
-
-    def process(self):
-
-        # render_all
-        for ent, (rend, pos) in self.world.get_components(Renderable, Position):
-            self.con.default_fg = rend.fg
-            self.con.default_bg = rend.bg
-            self.con.print_(
-                x=pos.x,
-                y=pos.y,
-                string=rend.char,
-                bg_blend=rend.bg_blend
-            )
-
-        # blit console
-        self.con.blit(
-            dest=self.root_console,
-            width=self.map_width,
-            height=self.map_height
-        )
-
-        # flush console
-        tcod.console_flush()
-
-        # clear_all
-        for ent, (rend, pos) in self.world.get_components(Renderable, Position):
-            self.con.print_(
-                x=pos.x,
-                y=pos.y,
-                string=' ',
-                bg_blend=rend.bg_blend
-            )
-
-
-# -----------------------------------------------------------------------------
-# Console Processor
-# -----------------------------------------------------------------------------
-
-class ConsoleProcessor(esper.Processor):
-    def __init__(self):
-        super().__init__()
-
-    def process(self):
-        for ent, event in self.world.get_component(Event):
-            if event.action.get('exit'):
-                sys.exit()
-
-            if event.action.get('fullscreen'):
-                tcod.console_set_fullscreen(not tcod.console_is_fullscreen())
-
-
-# -----------------------------------------------------------------------------
-# Entity
-# -----------------------------------------------------------------------------
-
-def player(x, y):
-    return (
-        Renderable('@'),
-        Position(x, y),
-        Event({}),
-    )
-
-
-# -----------------------------------------------------------------------------
-# Scene and Game Loop
-# -----------------------------------------------------------------------------
 
 class Scene:
 
@@ -263,16 +12,16 @@ class Scene:
 
     def on_start(self):
         processors = (
-            RenderProcessor(),
-            EventProcessor(),
-            MovementProcessor(),
-            ConsoleProcessor()
+            p.RenderProcessor(),
+            p.EventProcessor(),
+            p.MovementProcessor(),
+            p.ConsoleProcessor()
         )
         for num, proc in enumerate(processors):
             self.world.add_processor(proc, priority=num)
 
     def on_enter(self):
-        self.world.create_entity(*player(x=10, y=20))
+        self.world.create_entity(*e.player(x=10, y=20))
 
     def on_update(self):
         self.world.process()
