+++ src/processor/__init__.py
@@ -1,0 +1,4 @@
+from processor.render import RenderProcessor
+from processor.event import EventProcessor
+from processor.movement import MovementProcessor
+from processor.console import ConsoleProcessor
